Search improvement components improvement and refactors
