+++ PasswordMenager/frontend/src/utils/helpers.utils.ts
@@ -1,0 +1,5 @@
+export const branch = <T>(
+  condition: boolean,
+  isTrue: T,
+  fallBackValue: T
+): T => (condition ? isTrue : fallBackValue);
